refactor theta hash table make it reusable for both theta and tuple
